added docstrings and a smiple test case checking the number of subjects is correct when producing static representations
