+++ configs/tabularize.yaml
@@ -20,9 +20,6 @@ aggs:
   - "value/last"
   - "value/slope"
   - "value/intercept"
-  - "value/residual/sum"
-  - "value/residual/sum_sqd"
-numeric_value_impute_strategy: "drop"
 dynamic_threshold: 0.01
 numerical_value_threshold: 0.1
